+++ extensions/parquet/table/src/main/java/io/deephaven/parquet/table/transfer/DoubleTransfer.java
@@ -18,8 +18,9 @@ import org.jetbrains.annotations.NotNull;
 import java.nio.DoubleBuffer;
 
 final class DoubleTransfer extends PrimitiveTransfer<WritableDoubleChunk<Values>, DoubleBuffer> {
-    static DoubleTransfer create(@NotNull final ColumnSource<?> columnSource, @NotNull final RowSet tableRowSet, final int targetPageSize) {
-        final int maxValuesPerPage = Math.toIntExact(Math.min(tableRowSet.size(), targetPageSize / Double.BYTES));
+    static DoubleTransfer create(@NotNull final ColumnSource<?> columnSource, @NotNull final RowSet tableRowSet,
+                              final int targetPageSize) {
+        final int maxValuesPerPage = Math.toIntExact(Math.min(tableRowSet.size(), (long) (targetPageSize / Double.BYTES)));
         final double[] backingArray = new double[maxValuesPerPage];
         return new DoubleTransfer(
                 columnSource,
