+++ extensions/parquet/table/src/main/java/io/deephaven/parquet/table/transfer/IntTransfer.java
@@ -13,8 +13,9 @@ import org.jetbrains.annotations.NotNull;
 import java.nio.IntBuffer;
 
 final class IntTransfer extends PrimitiveTransfer<WritableIntChunk<Values>, IntBuffer> {
-    static IntTransfer create(@NotNull final ColumnSource<?> columnSource, @NotNull final RowSet tableRowSet, final int targetPageSize) {
-        final int maxValuesPerPage = Math.toIntExact(Math.min(tableRowSet.size(), targetPageSize / Integer.BYTES));
+    static IntTransfer create(@NotNull final ColumnSource<?> columnSource, @NotNull final RowSet tableRowSet,
+                              final int targetPageSize) {
+        final int maxValuesPerPage = Math.toIntExact(Math.min(tableRowSet.size(), (long) (targetPageSize / Integer.BYTES)));
         final int[] backingArray = new int[maxValuesPerPage];
         return new IntTransfer(
                 columnSource,
