+++ extensions/parquet/table/src/main/java/io/deephaven/parquet/table/transfer/LongTransfer.java
@@ -18,8 +18,9 @@ import org.jetbrains.annotations.NotNull;
 import java.nio.LongBuffer;
 
 final class LongTransfer extends PrimitiveTransfer<WritableLongChunk<Values>, LongBuffer> {
-    static LongTransfer create(@NotNull final ColumnSource<?> columnSource, @NotNull final RowSet tableRowSet, final int targetPageSize) {
-        final int maxValuesPerPage = Math.toIntExact(Math.min(tableRowSet.size(), targetPageSize / Long.BYTES));
+    static LongTransfer create(@NotNull final ColumnSource<?> columnSource, @NotNull final RowSet tableRowSet,
+                              final int targetPageSize) {
+        final int maxValuesPerPage = Math.toIntExact(Math.min(tableRowSet.size(), (long) (targetPageSize / Long.BYTES)));
         final long[] backingArray = new long[maxValuesPerPage];
         return new LongTransfer(
                 columnSource,
